requirements.txt
appdirs==1.4.3
Django==3.2.5
psycopg2-binary==2.8.6
djangorestframework==3.12.4
